Downgrade to 1.21.1. Spawn coordinates stored as doubles

+++ src/main/java/nl/getagripgal/persistentspawn/PersistentSpawnConfig.java
@@ -4,17 +4,17 @@ package nl.getagripgal.persistentspawn;
  * The spawn config as stored on disk.
  */
 public class PersistentSpawnConfig {
-    public int x;
-    public int y;
-    public int z;
+    public double x;
+    public double y;
+    public double z;
     public String dimension;
     public boolean enabled;
 
     public static PersistentSpawnConfig defaultConfig() {
         PersistentSpawnConfig config = new PersistentSpawnConfig();
-        config.x = 0;
-        config.y = 100;
-        config.z = 0;
+        config.x = 0.0;
+        config.y = 100.0;
+        config.z = 0.0;
         config.dimension = "minecraft:overworld";
         config.enabled = false;
         return config;
